chore(tests): match filename and classname
