+++ engine/tests/phpunit/integration/Elgg/Lib/PageOwnerIntegrationTest.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Elgg\Lib;
+
 use Elgg\IntegrationTestCase;
 use Elgg\Router\Route;
 
@@ -79,7 +81,7 @@ class PageOwnerIntegrationTest extends IntegrationTestCase {
 	}
 		
 	public function testSettingNegativeOwner() {
-		$this->expectException(InvalidArgumentException::class);
+		$this->expectException(\InvalidArgumentException::class);
 		_elgg_services()->pageOwner->setPageOwnerGuid(-1);
 	}
