+++ engine/tests/phpunit/unit/Elgg/Database/Clauses/AnnotationWhereClauseUnitTest.php
@@ -1,9 +1,7 @@
 <?php
 
-namespace Elgg\Integration;
+namespace Elgg\Database\Clauses;
 
-use Elgg\Database\Clauses\AccessWhereClause;
-use Elgg\Database\Clauses\AnnotationWhereClause;
 use Elgg\Database\QueryBuilder;
 use Elgg\Database\Select;
 use Elgg\UnitTestCase;
@@ -12,7 +10,7 @@ use Elgg\UnitTestCase;
  * @group QueryBuilder
  * @group QueryBuilderWhere
  */
-class AnnotationWhereClauseTest extends UnitTestCase {
+class AnnotationWhereClauseUnitTest extends UnitTestCase {
 
 	/**
 	 * @var QueryBuilder
