+++ engine/tests/phpunit/unit/Elgg/Database/Clauses/OrderByClauseUnitTest.php
@@ -1,19 +1,18 @@
 <?php
 
-namespace Elgg\Integration;
+namespace Elgg\Database\Clauses;
 
-use Elgg\Database\Clauses\SelectClause;
 use Elgg\Database\QueryBuilder;
 use Elgg\Database\Select;
 use Elgg\Project\Paths;
 use Elgg\UnitTestCase;
 
-require_once Paths::elgg() . 'engine/tests/test_files/database/clauses/CallableSelect.php';
+require_once Paths::elgg() . 'engine/tests/test_files/database/clauses/CallableOrderBy.php';
 
 /**
  * @group QueryBuilder
  */
-class SelectClauseTest extends UnitTestCase {
+class OrderByClauseUnitTest extends UnitTestCase {
 
 	/**
 	 * @var QueryBuilder
@@ -28,11 +27,11 @@ class SelectClauseTest extends UnitTestCase {
 
 	}
 
-	public function testBuildSelectClauseFromString() {
+	public function testBuildOrderByClauseFromString() {
 
-		$this->qb->select('alias.guid AS g');
+		$this->qb->orderBy('alias.guid', 'desc');
 
-		$query = new SelectClause('alias.guid AS g');
+		$query = new OrderByClause('alias.guid', 'desc');
 		$qb = Select::fromTable('entities', 'alias');
 		$qb->addClause($query);
 
@@ -40,26 +39,13 @@ class SelectClauseTest extends UnitTestCase {
 		$this->assertEquals($this->qb->getParameters(), $qb->getParameters());
 	}
 
-	public function testBuildSelectClauseFromClosure() {
+	public function testBuildOrderByClauseFromClosure() {
 
-		$this->qb->select('alias.guid AS g');
+		$this->qb->orderBy('alias.guid', 'asc');
 
-		$query = new SelectClause(function(QueryBuilder $qb) {
-			return 'alias.guid AS g';
-		});
-
-		$qb = Select::fromTable('entities', 'alias');
-		$qb->addClause($query);
-
-		$this->assertEquals($this->qb->getSQL(), $qb->getSQL());
-		$this->assertEquals($this->qb->getParameters(), $qb->getParameters());
-	}
-	
-	public function testBuildSelectClauseFromInvokableClass() {
-
-		$this->qb->select('alias.guid AS g');
-
-		$query = new SelectClause(\CallableSelect::class);
+		$query = new OrderByClause(function(QueryBuilder $qb) {
+			return 'alias.guid';
+		}, 'asc');
 
 		$qb = Select::fromTable('entities', 'alias');
 		$qb->addClause($query);
@@ -68,11 +54,24 @@ class SelectClauseTest extends UnitTestCase {
 		$this->assertEquals($this->qb->getParameters(), $qb->getParameters());
 	}
 	
-	public function testBuildSelectClauseFromStaticClassFunction() {
+	public function testBuildOrderByClauseFromInvokableClass() {
 
-		$this->qb->select('alias.guid AS g');
+		$this->qb->orderBy('alias.guid', 'asc');
 
-		$query = new SelectClause('\CallableSelect::callable');
+		$query = new OrderByClause(\CallableOrderBy::class, 'asc');
+
+		$qb = Select::fromTable('entities', 'alias');
+		$qb->addClause($query);
+
+		$this->assertEquals($this->qb->getSQL(), $qb->getSQL());
+		$this->assertEquals($this->qb->getParameters(), $qb->getParameters());
+	}
+	
+	public function testBuildOrderByClauseFromStaticClassFunction() {
+
+		$this->qb->orderBy('alias.guid', 'asc');
+
+		$query = new OrderByClause('\CallableOrderBy::callable', 'asc');
 
 		$qb = Select::fromTable('entities', 'alias');
 		$qb->addClause($query);
