+++ engine/tests/phpunit/unit/Elgg/DeprecationWrapperUnitTest.php
