+++ engine/tests/phpunit/unit/Elgg/I18n/TranslationLoadingUnitTestCase.php
@@ -1,13 +1,7 @@
 <?php
-/**
- *
- */
 
-namespace phpunit\unit\Elgg\I18n;
+namespace Elgg\I18n;
 
-
-use Elgg\Application;
-use Elgg\I18n\Translator;
 use Elgg\IntegratedUnitTestCase;
 
 class TranslationLoadingUnitTestCase extends IntegratedUnitTestCase {
